+++ src/params.cpp
@@ -11,6 +11,7 @@ using namespace std ;
 namespace params{
 
 char sSurface [255], sSpectraDir [255];
+char coordinateSystemHydro [10] = "milne";
 bool weakContribution ;
 bool rescatter ;
 bool shear ;
@@ -38,6 +39,20 @@ void readParams(char* filename)
 		sline >> parName >> parValue ;
 		if     (strcmp(parName,"surface")==0) strcpy(sSurface, parValue) ;
 		else if(strcmp(parName,"spectra_dir")==0) strcpy(sSpectraDir, parValue) ;
+		else if(strcmp(parName,"coordinateSystemHydro")==0) {
+			strcpy(coordinateSystemHydro, parValue);
+			for (unsigned int i=0; i<strlen(coordinateSystemHydro); i++) {
+				coordinateSystemHydro[i] = tolower(coordinateSystemHydro[i]);
+			}
+			if(strcmp(coordinateSystemHydro,"milne")!=0 &&
+				 strcmp(coordinateSystemHydro,"cartesian")!=0) {
+				throw std::invalid_argument(
+					std::string("Only 'Milne' or 'Cartesian' are supported as coordinate "
+					"systems for the hydro evolution. Provided was '") +
+					std::string(coordinateSystemHydro) + std::string("'. "
+					"Please update the config and try again."));
+			}
+		}
 		else if(strcmp(parName,"Nbins")==0) NBINS = atoi(parValue) ;
 		else if(strcmp(parName,"q_max")==0) QMAX = atof(parValue) ;
 		else if(strcmp(parName,"number_of_events")==0) NEVENTS = atoi(parValue) ;
@@ -57,9 +72,10 @@ void readParams(char* filename)
 
 void printParameters()
 {
-  cout << "====== parameters ======\n" ;
+  cout << "======= parameters =======\n" ;
   cout << "surface = " << sSurface << endl ;
   cout << "spectraDir = " << sSpectraDir << endl ;
+	cout << "coordinateSystemHydro = " << coordinateSystemHydro << endl;
   cout << "numberOfEvents = " << NEVENTS << endl ;
   cout << "isRescatter = " << rescatter << endl ;
   cout << "weakContribution = " << weakContribution << endl ;